requirements.txt
librosa
numpy
scipy
soundfile
mysql-connector-python
pandas
matplotlib
scikit-learn
flask
werkzeug